+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fix false positive when protocols requirements are satisfied in another file from the one that declares the conformance.
 - Fix redundant public accessibility analysis of enum associated value types.
+- Fix redundant public accessibility analysis of aliased types.
 
 ## 2.8.3 (2021-11-29)